+++ handler/live_mv/live_mv.go
@@ -26,7 +26,7 @@ func LiveMvStart(ctx *gin.Context) {
 	// err := json.Unmarshal([]byte(reqBody), &req)
 	// utils.CheckErr(err)
 
-	userId := ctx.GetInt("user_id")
+	userId := int32(ctx.GetInt("user_id"))
 	session := userdata.GetSession(ctx, userId)
 	defer session.Close()
 
@@ -42,7 +42,7 @@ func LiveMvSaveDeck(ctx *gin.Context) {
 	err := json.Unmarshal([]byte(reqBody), &req)
 	utils.CheckErr(err)
 
-	userId := ctx.GetInt("user_id")
+	userId := int32(ctx.GetInt("user_id"))
 	session := userdata.GetSession(ctx, userId)
 	defer session.Close()
